02-18(배열,열거타입)

+++ src/ch05homework/Exercise08.java
@@ -20,7 +20,7 @@ public class Exercise08 {
 			}
 		}
 		
-		avg = sum/count;
+		avg = (double)sum/count;
 		
 		System.out.println("sum " + sum);
 		System.out.println("avg "+ avg);
